- Reworked Scrolling examples to not need array index for buffering - Fixed browser incompatibiity for ESP8266 example

+++ examples/Parola_Scrolling_ML/Parola_Scrolling_ML.ino
@@ -130,7 +130,7 @@ void doUI(void)
 
 void readSerial(void)
 {
-  static int8_t	putIndex = -1;
+  static int16_t	putIndex = -1;
   char c;
 
   while (Serial.available())
